Replaced the function `CreateChildProcessAndExecuteSystemCommandWithIt` to `ExecuteSystemCommand`.
Because the child process of the most outer process had performed only
one in the following three tasks:
1. ExecuteHistoryCommand
2. ExecuteExclamationMarkCommand
3. CreateChildProcessAndExecuteSystemCommandWithIt
We could let it execute the system command directly as letting `execvp`
overwrite this child process would do no harm. By doing this, we saved
the need to fork an additional grandchild process.