Virtually re-wrote `history_supported_shell.c`.
Previously, all commands were executed in a separated child proccess.
Now, all commands except for system commands are executed in the parent
process (system commands remains being executed in a separate child
process). The reason for this is because in order to insert previous
system commands into the executed system commands -- the operation of
the exclamation mark command, this operation must be in the parent
process where the original `g_kExecutedSystemCommands` lies. Thus, we
must move the operation of the exclamation mark to the parent process.
For consistency, though not necessary, we also move the operation of the
history command to the parent process. Accordingly, only the system
commands will be executed in a separate child process.
Remark: This commit changes the whole code, so do not try to compare the
differences between this commit with the previous one, but read the
modified code directly instead.